refactor: co-locate contest tests and rename test_cases to fixtures
